Add a configuration option to configure max-items for large resources
This change adds a -m option which configures the --max-items API parameter for large
AWS resources. Currently, SNS topic subscriptions are limited to the default of 100
items. SNS topics can easily surpass 100,000 subscriptions which is too many to show
by default.

Since check 3.15 is confirming that subscribers exist - not what they actually are - it's
a waste to display all 100,000 entries.