Update check_extra7195
Fixed the check title from being incorrectly set to 'check7195' rather than 'extra7195'

For more context - I have a script that builds out 5 checklists based off of the check names returned from 'prowler -l', and then it spawns 5 subprocesses running prowler with each checklist. This was a need after running prowler using assumed roles in large environments, and my session expiring before prowler could be finished. The issue is that 'check7195' would be the 5th check in checlist_4.txt, and prowler would error out and not run that or subsequent checks.

I want to add the script as a merge request, but it is currently written in Fish shell syntax which is quite unique and wouldn't be that useful to others - so I plan on either changing it to bash or writing it in python at some point (bash would make sense since the whole tool is a bunch of bash scripts to begin with, and the syntax is nearly identical to Fish).