### Context
Updated the ECS check output to be more informative so that the user can more easily investigate the finding. It now outputs the types of secrets found and line number it was found on.

### Description

Created a secrets_string that contains the type of secret and line number for each secret detected, and adds it to the status_extended message

### License

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.