EC2 and AWS Lambda mock tests
Wrote tests for the EC2 check and aws lambda check. Found some errors so I fixed it and left comments.

I tried writing tests for ECS, but did not complete because I heard that ECS is not possible to create a fully functioning test since the check is not fully functioning