fix(kms): ensure kms GetKeyRotationStatus api call is attempted on all keys in self.keys regardless of an exception on a single key
When a KMS GetKeyRotationStatus api call on a single key fails with an exception (e.g. AccessDenied) subsequent keys in the
"self.keys" list should not be skipped because the api call on each key is independent of other keys in the list.
The current behaviour is that "key.rotation_enabled" of subsequent keys are not set as would be expected.