fix(kms): ensure kms GetKeyPolicy api call is attempted on all keys in self.keys
When a KMS GetKeyPolicy api call on a single key fails with an exception (e.g. AccessDenied) subsequent keys in the
"self.keys" list should not be skipped because the api call on each key is independent of other keys in the list.
The current behaviour is that "key.policy" of subsequent keys are not set as would be expected.
Also, "kms_key_not_public_accessible" check results of keys that failed GetKeyPolicy api call are incorrectly reported as "PASS",
a false positive result.